Added an additional BuildType of DateIncrement to the Version task.  It will set the buld number to be the number of days since the StartDate.  Useful if you want to track how long a project has been started.
Changed the way the RevisionType works when set to Increment and BuildType is set to DateIncrement.  It will start the first build of a new day at zero, then increment from there.  That way, if you do multiple builds to testers in a day, you have an easy way to track it.

Added Xsd annotation and the additional items to the intellisense schema.

Added additional unit tests to make sure nothing old or new will break.


git-svn-id: http://msbuildtasks.tigris.org/svn/msbuildtasks/trunk@233 299a232d-b705-0410-b782-f21f2f1e606a