Restore StandardErrorLoggingImportance pre-3.5 value
The default value for the StandardErrorLoggingImportance property on the
Microsoft.Build.Utilities.ToolTask class changed from High in .net 2.0
to Normal in 3.5 and up.

When the MSBuild.Community.Tasks project was migrated to 3.5 in pull
request #25 to fix the problem mentioned in issue #19, the change to
StandardErrorLoggingImportance was not accounted for, and several tasks
that depended on this to distinguish errors from regular messages in their
LogEventsFromTextOutput() methods stopped working correctly.

This commit adds StandardErrorLoggingImportance overrides to the
impacted tasks, restoring the pre-3.5 value of High for error messages.

[fixes #31]
[fixes #34]