Makes NuGet package add itself to projects.
https://docs.nuget.org/create/creating-and-publishing-a-package#from-a-convention-based-working-directory

A while back, NuGet added support for convertion-based folder called "build".  Any .targets file that is found in this folder will be imported into the project file this NuGet package is added to.

This resolves https://github.com/loresoft/msbuildtasks/issues/144, since there is no longer any on-install script that places files where they need to be.  Also, this pattern makes it so projects the NuGet package is added to get access to the all of the tasks immediately, without the user having to import the tasks themselves.