task documentation with examples
The readme on github lists a bunch of task names but provides no minimum
documentation on how to use each. This is a tweak to the readme tasks
table (and the xslt that appeared to be used for producing it) to link
each task to a documentation file for all tasks. The documentation is
generated from the XML comments in the C# code